requirements.txt
# pip install -r requirements.txt
vietocr~=0.3.11
uvicorn==0.17.6 # Newer version has deprecated debug in run()
fastapi~=0.88.0
pylibsrtp~=0.7.1
databases~=0.7.0
yolov5~=7.0.3
aiosqlite
python-multipart
Jinja2
shapely==1.7.1
# numpy>=1.18.5
# opencv-python>=4.1.2
# Pillow>=7.1.2
# PyYAML>=5.3.1
# requests>=2.23.0
# scipy>=1.4.1
# torch>=1.7.0
# torchvision>=0.8.1
# tqdm>=4.41.0
SQLAlchemy~=1.4.45
numpy~=1.24.0
pydantic~=1.10.2
Pillow~=9.3.0
opencv-python~=4.6.0.66